Move `error` submodules to separate files
